// front-end/src/types/EnvData.ts
export interface EnvData {
  country_code: string;
  frequency: string;
  measure_code: string;
  Measure: string;
  erosion_level: string;
  water_type: string;
  nutrient_type: string;
  Nutrients: string;
  unit: string;
  year: number;
  value: number;
  decimals: number;
  Decimals: string;
  status: string;
  UNIT_MULT: number;
  BASE_PER?: number;
}